DI-USB: - Fixed freeze when mounting a disc through overlay menu - Fixed mounting a disc the first time when inserting it in the disc slot
ES:
- Fixed menu patches when priiloader is installed on emulated nand
- Fixed booting Quadforce games with custom headers
- Fixed a timing issue that was introduced in r93 and broke support for some games (MHT, BBBP and probably more)
- Temp region feature works for titles also now
- Added tmd patch to region free titles "System Region Free Hack" is enabled
- Reduced memory usage
- Changed ES and DI logging options
- Moved some code to a better place to speed up game launching
- Added an option in overlay menu to monitor DI activity with slot LED
- Added an option in overlay menu to reverse the slot LED for DI activity monitor

FS-USB:
- Fixed compiler warnings
- Fixed use of external nandpath when no emulated nands are available in "/nands/" folder


Possible some other (WIP) stuff

+++ trunk/fs-usb/usb_os.c
@@ -17,9 +17,9 @@ int usb_os_init(void)
 //static u8* aligned_mem = 0;
 //static u8* aligned_base = 0;
 /* @todo hum.. not that nice.. */
-void*ehci_maligned(int size,int alignement,int crossing)
+void *ehci_maligned(int size, int alignement, int crossing)
 {
-	u32 addr = malloca( size, alignement );
+	u32 addr = (u32)malloca(size, alignement);
         //if (!aligned_mem )
         //{
         //        aligned_mem=aligned_base =  (void*)0x13890000;
@@ -38,10 +38,12 @@ void*ehci_maligned(int size,int alignement,int crossing)
         //memset((void*)addr,0,size);
 	return (void*)addr;
 }
+
 dma_addr_t ehci_virt_to_dma(void *a)
 {
 	return (dma_addr_t)virt_to_phys(a);
 }
+
 dma_addr_t ehci_dma_map_to(void *buf,size_t len)
 {
 	sync_after_write(buf, len);
@@ -67,7 +69,7 @@ void ehci_dma_unmap_from(dma_addr_t buf,size_t len)
 {
 	sync_before_read((void*)buf, len);
 }
-void ehci_dma_unmap_bidir(dma_addr_t buf,size_t len)
+void ehci_dma_unmap_bidir(dma_addr_t buf, size_t len)
 {
 	sync_before_read((void*)buf, len);
 }
